+++ changelog.md
@@ -3,6 +3,7 @@ Includes all important changes to the RPG adventure game.
 
 **v0.2 - Prerelease**
 - Added pseudocode v0.1
+- Updated game.py file to have numerical lists
 
 **v0.1 - Prerelease**
 - Added ReadMe Section describing the project.